Moved preprocess.py from package into scripts.
